chore: StyledNavigationContainer 위치 변경 	renamed:    src/navigation/components/styledBottomTabBar/StyledNavigationContainer.tsx -> src/navigation/components/styledNavigationContainer/StyledNavigationContainer.tsx 	renamed:    src/navigation/components/styledBottomTabBar/__test__/StyledNavigationContainer.test.tsx -> src/navigation/components/styledNavigationContainer/__test__/StyledNavigationContainer.test.tsx
