+++ src/navigation/components/styledNavigationContainer/StyledNavigationContainer.tsx
